// libs/core/dtos/enums/anime-source.dto.ts
/** Source dto. */
export enum AnimeSourceDto {
	FOUR_KOMA_MANGA = 'FOUR_KOMA_MANGA',
	BOOK = 'BOOK',
	CARD_GAME = 'CARD_GAME',
	GAME = 'GAME',
	LIGHT_NOVEL = 'LIGHT_NOVEL',
	MANGA = 'MANGA',
	MIXED_MEDIA = 'MIXED_MEDIA',
	MUSIC = 'MUSIC',
	NOVEL = 'NOVEL',
	ORIGINAL = 'ORIGINAL',
	PICTURE_BOOK = 'PICTURE_BOOK',
	RADIO = 'RADIO',
	VISUAL_NOVEL = 'VISUAL_NOVEL',
	WEB_MANGA = 'WEB_MANGA',
	WEB_NOVEL = 'WEB_NOVEL',
	OTHER = 'OTHER',
	UNKNOWN = 'UNKNOWN',
}