// libs/core/models/enums/anime-source.model.ts
/** Source. */
export enum AnimeSource {
	FourKomaManga = 'four koma manga',
	Book = 'book',
	CardGame = 'card game',
	Game = 'game',
	LightNovel = 'light novel',
	Manga = 'manga',
	MixedMedia = 'mixed media',
	Music = 'music',
	Novel = 'novel',
	Original = 'original',
	PictureBook = 'picture book',
	Radio = 'radio',
	VisualNovel = 'visual novel',
	WebManga = 'web manga',
	WebNovel = 'web novel',
	Other = 'other',
	Unknown = 'unknown',
}